use module files (#252)
* use module

* detect files to be restore automatically

* use esbuild

* drop ncc

* add comments

* refactor
